complete change from int to double, fix many minor bugs

+++ StoreManagement/Class/Client_Accessory_OrderedQuantity_Comment.cs
@@ -9,7 +9,7 @@ namespace StoreManagement.Class
     {
         public int AccessoryPK { get; set; }
 
-        public int OrderedQuantity { get; set; }
+        public double OrderedQuantity { get; set; }
 
         public string Comment { get; set; }
     }
