+++ StoreManagement/Controllers/Phase2/CountingItemController.cs
@@ -40,7 +40,7 @@ namespace StoreManagement.Controllers
             }
         }
 
-        public void updateCountingSession(int countingSessionPK, int countedQuantity)
+        public void updateCountingSession(int countingSessionPK, double countedQuantity)
         {
             try
             {
